Removed old calendar
